Fix Doctrine mappings
